+++ packages/plugin-packages/mashroom-portal-default-theme/src/backend/dev_server.ts
@@ -31,7 +31,6 @@ app.get('/', (req: Request, res: Response) => {
 
     const model: MashroomPortalPageRenderModel = {
         portalName: 'Test Portal',
-        portalBasePath: '/portal',
         siteBasePath: '/portal/web',
         lang: 'en',
         availableLanguages: ['en', 'de', 'fr'],