+++ packages/plugin-packages/mashroom-portal-demo-alternative-theme/src/plugin/dev_server.ts
@@ -26,7 +26,6 @@ app.get('/', (req: Request, res: Response) => {
 
     const model: MashroomPortalPageRenderModel = {
         portalName: 'Test Portal',
-        portalBasePath: '/portal',
         siteBasePath: '/portal/web',
         lang: 'en',
         availableLanguages: ['en', 'de', 'fr'],