+++ packages/plugin-packages/mashroom-portal/type-definitions/index.js
@@ -171,11 +171,10 @@ export type MashroomPortalPageContent = {
     +fullPageLoadRequired: ?boolean;
     +pageContent: string;
     +evalScript: string;
-};
+}
 
 export type MashroomPortalPageRenderModel = {
     +portalName: string,
-    +portalBasePath: string,
     +siteBasePath: string,
     +apiBasePath: string,
     +resourcesBasePath: ?string,
@@ -714,8 +713,21 @@ export interface MashroomPortalSiteService {
 
 export interface MashroomPortalPageService {
     /**
+     * Get current pageId
+     */
+    getCurrentPageId(): string;
+    /**
+     * Get the page friendlyUrl from given URL (e.g. /portal/web/test?x=1 -> /test)
+     */
+    getPageFriendlyUrl(pageUrl: string): string;
+    /**
+     * Find the pageId for given URL (can be a page friendlyUrl or a full URL as seen by the client).
+     */
+    getPageId(pageUrl: string): Promise<?string>;
+    /**
      * Get the content for given pageId.
-     * The request also calculates if a content replacement is possible without a full page load.
+     * It also calculates if the correct theme and all necessary page enhancements for the requested page
+     * are already loaded. Otherwise fullPageLoadRequired is going to be true and no content returned.
      */
     getPageContent(pageId: string): Promise<MashroomPortalPageContent>;
 }